+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed:
 - `Container` signature is now simplifed. It noq only receives elements using its constructor.
 - `Container` now inherits from `dict` to simplify its operations
+- `RSpace` now can transfer `dict` objects
 
 ## [0.0.27] - 2025-03-19
 ### Added: